+++ zxing-android-embedded/src/com/google/zxing/integration/android/IntentResult.java
@@ -28,21 +28,24 @@ public final class IntentResult {
     private final byte[] rawBytes;
     private final Integer orientation;
     private final String errorCorrectionLevel;
+    private final String barcodeImagePath;
 
     IntentResult() {
-        this(null, null, null, null, null);
+        this(null, null, null, null, null, null);
     }
 
     IntentResult(String contents,
                  String formatName,
                  byte[] rawBytes,
                  Integer orientation,
-                 String errorCorrectionLevel) {
+                 String errorCorrectionLevel,
+                 String barcodeImagePath) {
         this.contents = contents;
         this.formatName = formatName;
         this.rawBytes = rawBytes;
         this.orientation = orientation;
         this.errorCorrectionLevel = errorCorrectionLevel;
+        this.barcodeImagePath = barcodeImagePath;
     }
 
     /**
@@ -80,16 +83,21 @@ public final class IntentResult {
         return errorCorrectionLevel;
     }
 
-    @Override
-    public String toString() {
-        StringBuilder dialogText = new StringBuilder(100);
-        dialogText.append("Format: ").append(formatName).append('\n');
-        dialogText.append("Contents: ").append(contents).append('\n');
-        int rawBytesLength = rawBytes == null ? 0 : rawBytes.length;
-        dialogText.append("Raw bytes: (").append(rawBytesLength).append(" bytes)\n");
-        dialogText.append("Orientation: ").append(orientation).append('\n');
-        dialogText.append("EC level: ").append(errorCorrectionLevel).append('\n');
-        return dialogText.toString();
+    /**
+     * @return path to a temporary file containing the barcode image, if applicable, or null otherwise
+     */
+    public String getBarcodeImagePath() {
+        return barcodeImagePath;
     }
 
+    @Override
+    public String toString() {
+        int rawBytesLength = rawBytes == null ? 0 : rawBytes.length;
+        return "Format: " + formatName + '\n' +
+            "Contents: " + contents + '\n' +
+            "Raw bytes: (" + rawBytesLength + " bytes)\n" +
+            "Orientation: " + orientation + '\n' +
+            "EC level: " + errorCorrectionLevel + '\n' +
+            "Barcode image: " + barcodeImagePath + '\n';
+    }
 }