+++ zxing-android-embedded/src/com/journeyapps/barcodescanner/Size.java
@@ -1,5 +1,7 @@
 package com.journeyapps.barcodescanner;
 
+import android.support.annotation.NonNull;
+
 /**
  *
  */
@@ -34,6 +36,39 @@ public class Size implements Comparable<Size> {
     }
 
     /**
+     * Scales the dimensions so that it fits entirely inside the parent.One of width or height will
+     * fit exactly. Aspect ratio is preserved.
+     *
+     * @param into the parent to fit into
+     * @return the scaled size
+     */
+    public Size scaleFit(Size into) {
+        if(width * into.height >= into.width * height) {
+            // match width
+            return new Size(into.width, height * into.width / width);
+        } else {
+            // match height
+            return new Size(width * into.height / height, into.height);
+        }
+    }
+    /**
+     * Scales the size so that both dimensions will be greater than or equal to the corresponding
+     * dimension of the parent. One of width or height will fit exactly. Aspect ratio is preserved.
+     *
+     * @param into the parent to fit into
+     * @return the scaled size
+     */
+    public Size scaleCrop(Size into) {
+        if(width * into.height <= into.width * height) {
+            // match width
+            return new Size(into.width, height * into.width / width);
+        } else {
+            // match height
+            return new Size(width * into.height / height, into.height);
+        }
+    }
+
+    /**
      * Checks if both dimensions of the other size are at least as large as this size.
      *
      * @param other the size to compare with
@@ -47,7 +82,7 @@ public class Size implements Comparable<Size> {
      * Default sort order is ascending by size.
      */
     @Override
-    public int compareTo(Size other) {
+    public int compareTo(@NonNull Size other) {
         int aPixels = this.height * this.width;
         int bPixels = other.height * other.width;
         if (bPixels < aPixels) {
@@ -70,9 +105,7 @@ public class Size implements Comparable<Size> {
 
         Size size = (Size) o;
 
-        if (width != size.width) return false;
-        return height == size.height;
-
+        return width == size.width && height == size.height;
     }
 
     @Override