+++ zxing-android-embedded/src/com/journeyapps/barcodescanner/camera/CameraSettings.java
@@ -14,12 +14,18 @@ public class CameraSettings {
     private boolean continuousFocusEnabled = false;
     private boolean exposureEnabled = false;
     private boolean autoTorchEnabled = false;
+    private FocusMode focusMode = FocusMode.AUTO;
 
+    public enum FocusMode {
+        AUTO,
+        CONTINUOUS,
+        INFINITY,
+        MACRO
+    }
 
     public int getRequestedCameraId() {
         return requestedCameraId;
     }
-
 
     /**
      * Allows third party apps to specify the camera ID, rather than determine
@@ -98,6 +104,14 @@ public class CameraSettings {
 
     public void setAutoFocusEnabled(boolean autoFocusEnabled) {
         this.autoFocusEnabled = autoFocusEnabled;
+
+        if (autoFocusEnabled && continuousFocusEnabled) {
+            focusMode = FocusMode.CONTINUOUS;
+        } else if (autoFocusEnabled) {
+            focusMode = FocusMode.AUTO;
+        } else {
+            focusMode = null;
+        }
     }
 
     /**
@@ -111,6 +125,27 @@ public class CameraSettings {
 
     public void setContinuousFocusEnabled(boolean continuousFocusEnabled) {
         this.continuousFocusEnabled = continuousFocusEnabled;
+
+        if (continuousFocusEnabled) {
+            focusMode = FocusMode.CONTINUOUS;
+        } else if (autoFocusEnabled) {
+            focusMode = FocusMode.AUTO;
+        } else {
+            focusMode = null;
+        }
+    }
+
+    /**
+     * Default to FocusMode.AUTO.
+     *
+     * @return value of selected focus mode
+     */
+    public FocusMode getFocusMode() {
+        return focusMode;
+    }
+
+    public void setFocusMode(FocusMode focusMode) {
+        this.focusMode = focusMode;
     }
 
     /**