+++ CHANGES.md
@@ -1,3 +1,41 @@
+### 3.3.0 (2016-06-05)
+
+* Add an optional timeout to cancel scanning. (#161)
+* Rename CompoundBarcodeView to DecoratedBarcodeView.
+* Add more internal documentation (comments).
+
+### 3.2.0 (2016-02-06)
+
+* Improved preview scaling strategies, configurable between centerCrop, fitCenter, fitXY (#135)
+* Fix issues with Android 6 permission support (#123)
+* Fix camera initialization issues, specifically related to orientation changes (#133)
+* More control over focus mode (#112)
+* Keep drawing viewfinder frame after scanning / pausing (#134)
+* More control over torch state, and save the state on orientation change (#136)
+
+### 3.1.0 (2015-12-29)
+
+* Add support for Android 6 runtime permissions (Camera only).
+* Experimental support for using TextureView instead of SurfaceView.
+* Fix build issues with custom attributes.
+* Support library version 23+ is now a requirement.
+
+
+### 3.1.0 (2015-12-29)
+
+* Initial Android 6 permission supoprt
+
+### 3.0.3 (2015-08-16)
+
+* Fix for preview on Google Glass.
+* Make ViewfinderView extensible. (#75)
+* Add option to return image of barcode via Intents. (#72)
+
+### 3.0.2 (2015-07-21)
+
+* Add helper class for encoding barcodes.
+* Support custom layouts for CompoundBarcodeView.
+
 ### 3.0.1 (2015-06-10)
 
 * Fix auto-focus stopping when enabling the torch.